feat: refactor welcome flow by restructuring widget imports and adding login/register functionality
